requirements.txt
# Google Sheets API
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Testing
pytest==8.0.1
pytest-cov==4.1.0

# CLI
click==8.1.7

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
black==24.1.1
flake8==7.0.0
gspread==5.4.0  # or the latest version